+++ src/common/bar.c
@@ -1,5 +1,6 @@
+#include "common/bar.h"
+#include "common/color.h"
 #include "util/textModifier.h"
-#include "bar.h"
 
 // green, yellow, red: print the color on nth (0~9) block
 // set its value == 10 means the color will not be printed
@@ -15,37 +16,46 @@ void ffAppendPercentBar(FFstrbuf* buffer, uint8_t percent, uint8_t green, uint8_
     percent = (uint8_t)(percent + 5) / 10;
     assert(percent <= 10);
 
-    if(!instance.config.pipe)
-        ffStrbufAppendS(buffer, "\033[97m[ ");
-    else
-        ffStrbufAppendS(buffer, "[ ");
+    if(instance.config.barBorder)
+    {
+        if(!instance.config.pipe)
+            ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_WHITE "m[ ");
+        else
+            ffStrbufAppendS(buffer, "[ ");
+    }
 
     for (uint8_t i = 0; i < percent; ++i)
     {
         if(!instance.config.pipe)
         {
             if (i == green)
-                ffStrbufAppendS(buffer, "\033[32m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_GREEN "m");
             else if (i == yellow)
-                ffStrbufAppendS(buffer, "\033[93m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_YELLOW "m");
             else if (i == red)
-                ffStrbufAppendS(buffer, "\033[91m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_RED "m");
         }
-        ffStrbufAppendS(buffer, "■");
+        ffStrbufAppend(buffer, &instance.config.barCharElapsed);
     }
 
     if (percent < 10)
     {
         if(!instance.config.pipe)
-            ffStrbufAppendS(buffer, "\033[97m");
+            ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_WHITE "m");
         for (uint8_t i = percent; i < 10; ++i)
-            ffStrbufAppendS(buffer, "-");
+            ffStrbufAppend(buffer, &instance.config.barCharTotal);
+    }
+
+    if(instance.config.barBorder)
+    {
+        if(!instance.config.pipe)
+            ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_WHITE "m ]");
+        else
+            ffStrbufAppendS(buffer, " ]");
     }
 
     if(!instance.config.pipe)
-        ffStrbufAppendS(buffer, "\033[97m ]" FASTFETCH_TEXT_MODIFIER_RESET);
-    else
-        ffStrbufAppendS(buffer, " ]");
+        ffStrbufAppendS(buffer, FASTFETCH_TEXT_MODIFIER_RESET);
 }
 
 // if (green < yellow)
@@ -71,20 +81,20 @@ void ffAppendPercentNum(FFstrbuf* buffer, uint8_t percent, uint8_t green, uint8_
         if(green < yellow)
         {
             if (percent <= green)
-                ffStrbufAppendS(buffer, "\033[32m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_GREEN "m");
             else if (percent <= yellow)
-                ffStrbufAppendS(buffer, "\033[93m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_YELLOW "m");
             else
-                ffStrbufAppendS(buffer, "\033[91m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_RED "m");
         }
         else
         {
             if (percent >= green)
-                ffStrbufAppendS(buffer, "\033[32m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_GREEN "m");
             else if (percent >= yellow)
-                ffStrbufAppendS(buffer, "\033[93m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_YELLOW "m");
             else
-                ffStrbufAppendS(buffer, "\033[91m");
+                ffStrbufAppendS(buffer, "\e[" FF_COLOR_FG_LIGHT_RED "m");
         }
     }
     ffStrbufAppendF(buffer, "%u%%", (unsigned) percent);